Styling: Refactor lists
