+++ src/api/ThemeType.d.ts
@@ -2,13 +2,6 @@
  * Global type for a theme.
  */
 export type ThemeType = {
-  list: {
-    header: {
-      background: string;
-      color: string;
-      shadow: string;
-    };
-  };
   groups: {
     header: {
       color: string;
