+++ src/styles/tab.ts
